+++ test/helpers/command-tester.js
