+++ test/helpers/read-dependencies-mock.js
