+++ test/lib/payee-calculator-test.js
@@ -1,6 +1,6 @@
 /* global describe, it, beforeEach, afterEach */
 var expect = require('chai').expect
-var sustainFixture = require('../lib/sustain-fixture-manager')
+var sustainFixture = require('../helpers/sustain-fixture-manager')
 var PayeeCalculator = require('../../lib/payee-calculator')
 
 process.chdir(__dirname + '/..')
